Moved more services around
